Add all tests
